refactor(web/applications): merge lib and common (boas-740)
