+++ apps/web/src/server/applications/pages/case/applications-case.controller.js
@@ -1,4 +1,4 @@
-import { publishCase } from '../../lib/services/case.service.js';
+import { publishCase } from '../../common/services/case.service.js';
 
 /** @typedef {import('../../applications.types').Case} Case */
 /** @typedef {import('@pins/express').ValidationErrors} ValidationErrors */
