+++ apps/web/src/server/applications/pages/create-new-case/applicant/applications-create-applicant.service.js
@@ -1,11 +1,11 @@
 import {
 	getSessionApplicantInfoTypes,
 	getSessionCaseHasNeverBeenResumed
-} from '../../../lib/services/session.service.js';
+} from '../../../common/services/session.service.js';
 
 /** @typedef {import('../../../applications.types').OptionsItem} OptionsItem */
 /** @typedef {import('../../../applications.types').Applicant} Applicant */
-/** @typedef {import('../../../lib/services/session.service.js').SessionWithApplicationsCreateApplicantInfoTypes} SessionWithApplicationsCreateApplicantInfoTypes */
+/** @typedef {import('../../../common/services/session.service.js').SessionWithApplicationsCreateApplicantInfoTypes} SessionWithApplicationsCreateApplicantInfoTypes */
 
 /**
  * Get the list of types of applicant info
