+++ apps/web/src/server/applications/pages/create-new-case/case/applications-create-case.controller.js
@@ -1,4 +1,5 @@
 import pino from '../../../../lib/logger.js';
+import { handleErrors } from '../../../common/components/error-handler/error-handler.component.js';
 import {
 	caseGeographicalInformationData,
 	caseGeographicalInformationDataUpdate,
@@ -14,13 +15,12 @@ import {
 	caseTeamEmailDataUpdate,
 	caseZoomLevelData,
 	caseZoomLevelDataUpdate
-} from '../../../components/form/form-case-components.controller.js';
-import { handleErrors } from '../../../lib/controllers/errors.controller.js';
+} from '../../../common/components/form/form-case.component.js';
 import {
 	destroySessionCaseSectorName,
 	getSessionCaseHasNeverBeenResumed,
 	setSessionCaseSectorName
-} from '../../../lib/services/session.service.js';
+} from '../../../common/services/session.service.js';
 
 /** @typedef {import('./applications-create-case.types.js').ApplicationsCreateCaseNameProps} ApplicationsCreateCaseNameProps */
 /** @typedef {import('./applications-create-case.types.js').ApplicationsCreateCaseNameBody} ApplicationsCreateCaseNameBody */
